+++ src/main/java/com/souf/soufwebsite/domain/application/controller/StudentApplicationApiSpecification.java
@@ -12,19 +12,17 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
+@Tag(name = "Application", description = "지원 관련 API")
 public interface StudentApplicationApiSpecification {
 
-    @Tag(name = "Application", description = "지원 관련 API")
     @Operation(summary = "지원하기", description = "특정 공고문에 지원합니다.")
     @PostMapping("/{recruitId}/apply")
     SuccessResponse<?> apply(@PathVariable Long recruitId);
 
-    @Tag(name = "Application", description = "지원 관련 API")
     @Operation(summary = "지원 취소", description = "특정 공고문에 대한 지원을 취소합니다.")
     @DeleteMapping("/{recruitId}/apply")
     SuccessResponse<?> deleteApplication(@PathVariable Long recruitId);
 
-    @Tag(name = "Application", description = "지원 관련 API")
     @Operation(summary = "내 지원 리스트 조회", description = "사용자 본인이 지원한 공고문 리스트를 조회합니다.")
     @GetMapping("/my")
     SuccessResponse<Page<MyApplicationResDto>> getMyApplications(@PageableDefault(size = 10) Pageable pageable);
